Dhosokaw 01

+++ project/src/draw_info.c
@@ -6,7 +6,7 @@
 /*   By: hosokawa <hosokawa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/11 15:31:32 by hosokawa          #+#    #+#             */
-/*   Updated: 2025/01/11 15:46:15 by hosokawa         ###   ########.fr       */
+/*   Updated: 2025/01/29 13:48:24 by hosokawa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ void	get_draw_texture_info(t_draw *draw)
 	int	h;
 
 	h = HEIGHT;
-	draw->step = 1.0 * TEXHEIGHT / draw->lineHeight;
+	draw->step = 1.0 * TEXHEIGHT / (double)draw->lineHeight;
 	draw->texPos = (draw->drawStart - h / 2 + draw->lineHeight / 2)
 		* draw->step;
 }
